% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Real-networks.R
\name{convert_gmresults_to_df}
\alias{convert_gmresults_to_df}
\title{Compute statistics from shuffle_gm_sim results}
\usage{
convert_gmresults_to_df(errors_list, p_grid, data = NULL)
}
\arguments{
\item{errors_list}{Generated from \code{\link{shuffle_gm_sim}}}

\item{p_grid}{Grid of probabilities used for above.}

\item{data}{name of the data}
}
\description{
Compute statistics from shuffle_gm_sim results
}
