% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GM_functions.R
\name{corrupting_channel}
\alias{corrupting_channel}
\title{Generate corrupted graph from A with probability p}
\usage{
corrupting_channel(A, p)
}
\arguments{
\item{A}{adjacency matrix}

\item{p}{corrupting probability}
}
\value{
Corrupted graph
}
\description{
Each edge/non-edge is flipped with probability p.
}
