% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Real-networks.R
\name{get_disag_stat}
\alias{get_disag_stat}
\alias{par_run_all_disag}
\title{Disagreements simulation}
\usage{
get_disag_stat(graph, cl = NULL, m = 500)

par_run_all_disag(nclust = 7, ask = TRUE)
}
\arguments{
\item{graph}{iGraph object or string corresponding to graph data in package}

\item{cl}{Cluster to run the shuffling on.}

\item{m}{Number of shuffles to perform to compute statistics}

\item{nclust}{Number of parallel clusters to create via
parallel::makeCluster}

\item{ask}{Whether to warn about long run times.}
}
\value{
Tibble with 7 columns
 K: number nodes permuted, on a logarithmic grid
 avedis: average distance between original and permuted graph
 sd: standard deviation of distance,
 min: minimum distance
 pthresh: estimated bound on error tolerance for graph matching
 data: name of data if graph is a string
 n: number of nodes in provided graph
}
\description{
Simulations to test for the difficulty to match
 a graph by randomly permuting nodes. get_disag_stat
 runs on one graph and par_run_all_disag runs the
 simulation in the manuscript in parallel.
}
\section{Functions}{
\itemize{
\item \code{par_run_all_disag}: 
}}

\examples{
get_disag_stat("karate", m = 10)
}
