% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Real-networks.R
\name{gmmle_data}
\alias{gmmle_data}
\title{Tibble of data used in the manuscript}
\usage{
gmmle_data()
}
\value{
A tibble with 3 columns, one for the name
 which corresponds to the variable name for the
 included data, one for number of nodes, and one
 for a formal name used for plotting.
 Names are factors ordered by number of nodes.
}
\description{
Datasets are Karate, C.elegans, 
  Politcal blogs, Protein interactions, and arXiv
  citations.
}
