% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Real-networks.R
\name{graph_from_name}
\alias{graph_from_name}
\title{Gets the igraph object for data from name}
\usage{
graph_from_name(gname)
}
\arguments{
\item{gname}{Name of graph.}
}
\value{
Corresponding igraph object
}
\description{
Gets the igraph object for data from name
}
\examples{
 graph_from_name("karate")
}
\seealso{
\code{\link{gmmle_data}}
}
