% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GM_functions.R
\name{kshuffle_mean_sd}
\alias{kshuffle_mean_sd}
\title{Statistics for errors induced by shuffling}
\usage{
kshuffle_mean_sd(k, G, m = 1000)
}
\arguments{
\item{k}{number of shuffled vertices}

\item{G}{graph object (igraph)}

\item{m}{number of random permutations}
}
\value{
A vector of length 4. The first three elements
 are the mean, sd, and minimum for the number of edge
 disagreements. The 4th, pmax is a theoretical bound
 on maximum error for which graph matching will be
 possible.
}
\description{
Calculate edge disagreements and probability bounds
  after k vertices are randomly shuffled in a graph G
}
\examples{
 kshuffle_mean_sd(10, karate, m = 10)
}
