% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GM_functions.R
\name{local_gm_errs}
\alias{local_gm_errs}
\title{Vertex matching errors for a corrupted graph}
\usage{
local_gm_errs(A, p, seed = 0)
}
\arguments{
\item{A}{adjacency}

\item{p}{corrupting probability}

\item{seed}{random seed}
}
\value{
list of diagonal of match and the proportion of errors
}
\description{
The graph A is corrupted with probability p and then
 the corrupted graph and the original graphs are matched
 with initialization at the identity.
}
\examples{
local_gm_errs(karate[], .05)

}
