% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{make_er_plot}
\alias{make_er_plot}
\title{Simulation 1: ER models}
\usage{
make_er_plot(data = NULL)
}
\arguments{
\item{data}{List of data frames created by simulation_instance_ER.
NULL for data from the manuscript.}
}
\description{
Make plot from simulations generated by simulation_instance_ER.
}
\examples{
 make_er_plot()

}
