% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{make_model_compare_plot}
\alias{make_model_compare_plot}
\title{Simulation 3: comparing different models for degree = 5 and 50}
\usage{
make_model_compare_plot(data5 = NULL, data50 = NULL)
}
\arguments{
\item{data5}{List of data frames created by simulation_instance for d=5.
NULL for data from the manuscript.}

\item{data50}{List of data frames created by simulation_instance for d=50.
NULL for data from the manuscript.}
}
\description{
Make plot from simulations generated by simulation_instance
}
\examples{
 make_model_compare_plot()

}
