% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{make_ws_plot}
\alias{make_ws_plot}
\title{Simulation 2: WS models}
\usage{
make_ws_plot(data = NULL)
}
\arguments{
\item{data}{List of data frames created by simulation_instance_WS.
NULL for data from the manuscript.}
}
\description{
Make plot from simulations generated by simulation_instance_WS.
}
\examples{
 make_ws_plot()

}
