% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Real-networks.R
\name{plot_matching_errors}
\alias{plot_matching_errors}
\title{Plot graph matching errors as a function or error rate}
\usage{
plot_matching_errors(
  data_df = NULL,
  p_grid = 10^(-seq(0.5, 4, length.out = 8))
)
}
\arguments{
\item{data_df}{A tibble with columns name, n, and
formal_name as in \code{\link{gmmle_data}} and a column
results which has output from \code{\link{shuffle_gm_sim}}
for each graph.}

\item{p_grid}{Grid of error probabilities used to generate results}
}
\description{
Plot graph matching errors as a function or error rate
}
\examples{
plot_matching_errors()

}
