% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Real-networks.R
\name{plot_prob_bound_norm_disag}
\alias{plot_prob_bound_norm_disag}
\title{Plot probability bounds and normalized mean disag}
\usage{
plot_prob_bound_norm_disag(data = NULL)
}
\arguments{
\item{data}{Should be either NULL for data used in the manuscript
or something like the output of \code{\link{par_run_all_disag}}}
}
\description{
Plot showing the predicted probability bounds
and the normalized mean degree as a function of the 
fraction of shuffled vertices.
}
\examples{
 plot_prob_bound_norm_disag()

}
