% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{realnet_error_df}
\alias{realnet_error_df}
\alias{disag_data}
\title{Real data simulations results.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 4 columns.

An object of class \code{data.frame} with 106 rows and 4 columns.
}
\source{
Generated by code from this package.
}
\usage{
realnet_error_df

disag_data
}
\description{
Results from the different simulations to remake the plots. 
 realnet_error_df has results from corrupting and graph matching.
 disag_data has results from permuting and computing disagreements.
}
\keyword{datasets}
