% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{run_all_simulations}
\alias{run_all_simulations}
\title{Run all simulations from the paper.}
\usage{
run_all_simulations(ask = TRUE)
}
\arguments{
\item{ask}{Ask about running big simulations.}
}
\description{
This replicates the results from the paper.
}
\seealso{
\code{\link{simulation_instance}},
 \code{\link{simulation_instance_ER}},
 \code{\link{simulation_instance_WS}}
}
