% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parsim_er}
\alias{parsim_er}
\alias{parsim_ws}
\alias{parsim5}
\alias{parsim50}
\title{Simulation results from the paper.}
\format{
An object of class \code{list} of length 35.

An object of class \code{list} of length 35.

An object of class \code{list} of length 35.

An object of class \code{list} of length 35.
}
\source{
Generated by code from this package.
}
\usage{
parsim_er

parsim_ws

parsim5

parsim50
}
\description{
Results from the different simulations to remake the plots. 
 Each is a list of data frames.
}
\keyword{datasets}
